:truck: Move utils to contrib package
