+++ src/openpersonen/contrib/stufbg/converters/partner_historie.py
@@ -2,7 +2,7 @@ from django.conf import settings
 
 import xmltodict
 
-from openpersonen.api.utils import convert_empty_instances
+from openpersonen.contrib.utils import convert_empty_instances
 
 
 def convert_response_to_partner_historie_dict(response):
